Crear TODO con usuario

+++ src/app/api/todos/route.ts
@@ -1,6 +1,7 @@
 import { NextResponse } from 'next/server';
 import prisma from '@/lib/prisma';
 import { boolean, object, string } from 'yup';
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 
 export async function GET( req: Request ) {
     const { searchParams } = new URL( req.url );
@@ -29,10 +30,15 @@ const postSchema = object( {
 } );
 
 export async function POST( req: Request ) {
+    const user = await getUserSessionServer();
+    if ( !user ) {
+        return NextResponse.json( { error: 'Unauthorized' }, { status: 401 } );
+    }
+
     try {
         const { description, completed } = await postSchema.validate( await req.json() );
 
-        const data = { completed, description };
+        const data = { completed, description, userId: user.id };
 
         const todo = await prisma.todo.create( { data } );
 
@@ -43,8 +49,13 @@ export async function POST( req: Request ) {
 }
 
 export async function DELETE( req: Request ) {
+    const user = await getUserSessionServer();
+    if ( !user ) {
+        return NextResponse.json( { error: 'Unauthorized' }, { status: 401 } );
+    }
+
     try {
-        await prisma.todo.deleteMany( { where: { completed: true } } );
+        await prisma.todo.deleteMany( { where: { completed: true, userId: user.id } } );
     } catch ( error ) {
         return NextResponse.json( error, { status: 400 } );
     }
